Optimize FastActivator.CreateInstanceDynamically()
The lookup of the creator has been improved significantly by using a specialized lookup tree instead of two cascaded dictionaries. The invocation of the creator is now done by a dynamically compiled lambda expression instead of using 'dynamic' variables to call to the creator. This furthermore eliminates the need for the System.CSharp package.